pyproject.toml
[tool.poetry]
name = "flaskboardx"
version = "0.1.0"
description = ""
authors = ["Hichem Dridi <dridi.hichem18@gmail.com>"]
readme = "README.md"
packages = [{include = "flaskboardx", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
click = "^8.1.7"
python-dotenv = "^1.0.0"
werkzeug = "^3.0.1"
flake8 = "^7.0.0"
pylint = "^3.0.3"
black = "^23.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
